% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_photon.R
\name{get_instance}
\alias{get_instance}
\alias{get_photon_url}
\title{Photon utilities}
\usage{
get_instance()

get_photon_url()
}
\value{
\code{get_instance} returns a R6 object of class \code{photon}.
\code{get_photon_url()} returns a URL string.
}
\description{
Utilities to manage photon instances. These functions operate on mounted
photon instances which can be initialized using \code{\link{new_photon}}.

\itemize{
\item{\code{get_instance()} retrieves the active photon instance.}
\item{\code{get_photon_url()} retrieves the photon URL to send requests.}
}
}
\examples{
# make a new photon instance
new_photon()

# retrieve it from the cache
get_instance()

# get the server url
get_photon_url()
}
