% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java.R
\name{has_java}
\alias{has_java}
\title{Is Java installed?}
\usage{
has_java(version = NULL)
}
\arguments{
\item{version}{Character string specifying the minimum version of Java.
If the installed Java version is lower than this, returns \code{FALSE}.
If \code{NULL}, only checks if any kind of Java is installed on the
system.}
}
\value{
A logical vector of length 1.
}
\description{
Utility function to check if Java is installed and if it has the right
version.
}
\examples{
has_java() # Is Java installed?
has_java("11") # Is Java > 11 installed?
}
