% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{import_options}
\alias{import_options}
\title{Nominatim import options}
\usage{
import_options(
  host = NULL,
  port = NULL,
  database = NULL,
  user = NULL,
  password = NULL,
  import = FALSE
)
}
\arguments{
\item{host}{Postgres host of the Nominatim database. Defaults to
\code{127.0.01}.}

\item{port}{Postgres port of the Nominatim database. Defaults to
\code{5432}.}

\item{database}{Postgres database name. Defaults to \code{nominatim}.}

\item{user}{Postgres user. Defaults to \code{nominatim}.}

\item{password}{Postgres database password. Defaults to \code{""}.}

\item{import}{If \code{TRUE}, uses the \code{-nominatim-import} option in
any case, even if no other option is specified. If \code{FALSE}, returns
\code{NULL}, indicating that photon should not import a Nominatim database.}
}
\value{
A character string of formatted command line options. Should only
be used in accordance with \code{$start()} from the \code{\link{photon_local}}
class. If all parameters are \code{NULL} and \code{import = FALSE}, returns
\code{NULL}.
}
\description{
Photon options to set when importing a Nominatim database. Can be used
together with the \code{$start()} method of \code{\link{photon_local}}.
}
\examples{
if (FALSE) {
dir <- file.path(tempdir(), "photon")
photon <- new_photon(dir, nominatim = TRUE)

# do not import
photon$start(import = import_options())

# import with defaults
photon$start(import = import_options(import = TRUE))

# import nominatim on non-default port
photon$start(import = import_options(port = 14149))

# import nominatim with non-default postgres user/password
photon$start(import = import_options(user = "harry", password = "securepw"))
}
}
