% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{photon}
\alias{photon}
\title{Local photon instance}
\description{
This R6 class is used to initialize and manage local photon instances.
It can download and setup the Java, the photon executable, and the necessary
Elasticsearch search index. It can start, stop, and query the status of the
photon instance. It is also the basis for geocoding requests at it is used
to retrieve the URL for geocoding.
}
\examples{
if (FALSE) {
# start a new instance using a Monaco extract
photon <- new_photon(path = tempdir(), country = "Monaco")

# start a new instance with an older photon version
photon <- new_photon(path = tempdir(), photon_version = "0.4.1")

# start a new instance with a specific java version
photon <- new_photon(path = tempdir(), java_version = 17)
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to the directory where the photon instance is stored.}

\item{\code{proc}}{\code{\link[processx]{process}} object that handles the
external process running photon.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-photon-new}{\code{photon$new()}}
\item \href{#method-photon-purge}{\code{photon$purge()}}
\item \href{#method-photon-start}{\code{photon$start()}}
\item \href{#method-photon-stop}{\code{photon$stop()}}
\item \href{#method-photon-is_running}{\code{photon$is_running()}}
\item \href{#method-photon-get_url}{\code{photon$get_url()}}
\item \href{#method-photon-clone}{\code{photon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-photon-new"></a>}}
\if{latex}{\out{\hypertarget{method-photon-new}{}}}
\subsection{Method \code{new()}}{
Initialize a local photon instance. If necessary, downloads the photon
executable, the search index, and Java.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{photon$new(
  path = "./photon",
  photon_version = NULL,
  java_version = 22,
  ...,
  country = NULL,
  date = "latest",
  exact = FALSE,
  consent = FALSE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to a directory where the photon executable and data
should be stored. Defaults to a directory "photon" in the current
working directory.}

\item{\code{photon_version}}{Version of photon to be used. A list of all
releases can be found here: \url{https://github.com/komoot/photon/releases/}.}

\item{\code{java_version}}{Java version to be used. See \code{\link[rJavaEnv]{use_java}}
for available versions.}

\item{\code{...}}{Further arguments passed to \code{\link[rJavaEnv]{use_java}}.}

\item{\code{country}}{Character string that can be identified by
\code{\link[countrycode]{countryname}} as a country. An extract for this
country will be downloaded. If \code{NULL}, downloads a global search index.}

\item{\code{date}}{Character string or date-time object used to specify the creation
date of the search index. If \code{"latest"}, will download the file tagged
with "latest". If a character string, the value should be parseable by
\code{\link{as.POSIXct}}. If \code{exact = FALSE}, the input value is
compared to all available dates and the closest date will be selected.
Otherwise, a file will be selected that exactly matches the input to
\code{date}.}

\item{\code{exact}}{If \code{TRUE}, exactly matches the \code{date}. Otherwise,
selects the date with lowest difference to the \code{date} parameter.}

\item{\code{consent}}{Whether to provide consent to \code{rJavaEnv} to setup
Java. See \code{\link[rJavaEnv]{rje_consent}}.}

\item{\code{quiet}}{If \code{TRUE}, suppresses all informative messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-photon-purge"></a>}}
\if{latex}{\out{\hypertarget{method-photon-purge}{}}}
\subsection{Method \code{purge()}}{
Kill the photon process and remove the directory. Useful to get rid
of an instance entirely.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{photon$purge()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-photon-start"></a>}}
\if{latex}{\out{\hypertarget{method-photon-start}{}}}
\subsection{Method \code{start()}}{
Start a local instance of the Photon geocoder. Runs the jar executable
located in the instance directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{photon$start(
  min_ram = 5,
  max_ram = 10,
  host = "0.0.0.0",
  port = "2322",
  java_options = NULL,
  photon_options = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_ram}}{Initial RAM to be allocated to the Java process
(\code{-Xms} flag).}

\item{\code{max_ram}}{Maximum RAM to be allocated to the Java process
(\code{-Xmx} flag)}

\item{\code{host}}{Character string of the host name that the geocoder should
be opened on.}

\item{\code{port}}{Port that the geocoder should listen to.}

\item{\code{java_options}}{List of further flags passed on to the \code{java}
command.}

\item{\code{photon_options}}{List of further flags passed on to the photon
jar in the java command.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-photon-stop"></a>}}
\if{latex}{\out{\hypertarget{method-photon-stop}{}}}
\subsection{Method \code{stop()}}{
Kills the running photon process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{photon$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-photon-is_running"></a>}}
\if{latex}{\out{\hypertarget{method-photon-is_running}{}}}
\subsection{Method \code{is_running()}}{
Checks whether the photon instance is running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{photon$is_running()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-photon-get_url"></a>}}
\if{latex}{\out{\hypertarget{method-photon-get_url}{}}}
\subsection{Method \code{get_url()}}{
Constructs the URL that geocoding requests should be sent to.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{photon$get_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-photon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-photon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{photon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
