% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_build.R
\name{run_build}
\alias{run_build}
\title{The main photon function}
\usage{
run_build(type = c("run", "build"), electron_shiny_path, app_path, r_path = R.home())
}
\arguments{
\item{type}{Choose whether to run or build the app. I suggest run it before you build it.}

\item{electron_shiny_path}{Location of your electron project, obtained by cloning TBD.}

\item{app_path}{Location of a folder containing your app files. These will be copied into {electron_shiny_path}/app.}

\item{r_path}{Location of the R installation to use for the app. These will be copied into {electron_shiny_path}/R.}
}
\description{
The main photon function
}
