% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-photon.R
\name{download_photon}
\alias{download_photon}
\title{Download photon}
\usage{
download_photon(path = ".", version = NULL, quiet = FALSE)
}
\arguments{
\item{path}{Path to a directory to store the exectutable.}

\item{version}{Version tag of the photon release.}
}
\value{
Character string giving the path to the downloaded file.
}
\description{
Download the photon executable from GitHub.
}
\examples{
if (FALSE) {
download_photon(tempdir(), version = "0.4.1")
}
}
