% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode Locations With Photon API}
\usage{
geocode(location, limit = NULL, key = NULL, value = NULL, lang = NULL,
  server = NULL, quiet = FALSE)
}
\arguments{
\item{location}{Location(s) to geocode, a character vector.}

\item{limit}{Maximum number of returned results.}

\item{key}{Photon allows filtering based on OSM tags and values (see photon
documentation).}

\item{value}{Photon allows filtering based on OSM tags and values (see photon
documentation).}

\item{lang}{Photon is multilingual, this parameter allows to specify a
language.}

\item{server}{Select which API is used. The default API is the photon public
API (see Details), but you can use your own photon API.}

\item{quiet}{Should the function print the currently processed input ?}
}
\value{
A data.frame with the following fields is returned :\cr
location, name, housenumber, street, postcode, city, state, country, osm_key,
osm_value,  lon,  lat, msg.\cr
\code{location} is the original searched location.\cr
\code{lon} and \code{lat} are longitude and latitude (WGS84).\cr
\code{msg} is filled with the "Not found" text string if no results are found for a
location.
}
\description{
This function geocode locations using the photon API.
}
\details{
Terms of use of the photon public API: \cr
"You can use the API for your
project, but please be fair - extensive usage will be throttled. We do not
guarantee for the availability and usage might be subject of change in the
future. Have fun with photon and make OSM grow!"
}
\examples{
loc <- c("19 rue Michel Bakounine, 29600 Morlaix, France",
         "5 rue Proudhon, 34000 Montpellier, France",
         "2 Emma Goldmanweg, Tilburg, Netherlands",
         "36 Strada Panait Istrati, Bucarest, Romania")
geocode(loc, limit = 1, key = "place")
}
\references{
Photon web site \url{http://photon.komoot.de/}\cr
Photon source code \url{https://github.com/komoot/photon}
}

