% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage.R
\name{new_photon}
\alias{new_photon}
\title{Initialize a photon instance}
\usage{
new_photon(path = NULL, url = NULL, ...)
}
\arguments{
\item{path}{Path to a directory where the photon executable and data
should be stored. Defaults to a directory "photon" in the current
working directory. If \code{NULL}, a remote instance is set up based on
the \code{url} parameter.}

\item{url}{If \code{type = "public"}, specifies the URL that geocoding
requests are sent to. If \code{NULL} and \code{path} is also \code{NULL},
connects to the public API under \url{photon.komoot.io}.}

\item{...}{Arguments passed to \code{\link{photon}$new()}.}
}
\value{
An R6 object of class \code{photon}.
}
\description{
Initialize a photon instance by creating a new photon object. This object
is stored in the R session and can be used to perform geocoding requests.

Instances can either local or remote. Remote instances require nothing more
than a URL that geocoding requests are sent to. Local instances require the
setup of the photon executable, an Elasticsearch index, and Java. See
\code{\link{photon}} for details.
}
\examples{
# connect to public API
photon <- new_photon()

\dontrun{
# connect to arbitrary server
photon <- new_photon(url = "photonserver.org")

# set up a local instance
photon <- new_photon(path = tempdir())}
}
