% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photon-package.R
\docType{package}
\name{photon-package}
\alias{photon-package}
\title{photon: Interface to the 'Photon' geocoding API}
\description{
Features unstructured, structured and reverse geocoding using the 'Photon' geocoding API. Facilitates the setup of local Photon instances to enable offline geocoding.
}
\author{
\strong{Maintainer}: Jonas Lieth [copyright holder]

}
\keyword{internal}
