% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\docType{package}
\name{photon}
\alias{photon}
\alias{photon-package}
\title{Geocode Locations With Photon API}
\description{
Photon is an open source geocoder built for OpenStreetMap data
and based on elasticsearch. This package allows to query a photon API and get
the results in a data frame.
}
