% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge_java}
\alias{purge_java}
\title{Purge Java processes}
\usage{
purge_java(pids = NULL, consent = FALSE)
}
\arguments{
\item{pids}{PIDs to kill. The PIDs should be Java processes. If \code{NULL},
tries to kill all Java processes.}

\item{consent}{If \code{FALSE}, asks for consent before killing the
processes. Defaults to \code{FALSE}.}
}
\value{
An integer vector of the \code{pkill} / \code{Taskkill} status
codes or \code{NULL} if not running Java processes are found.
}
\description{
Kill all or selected running Java processes. This function is useful to
stop Photon instances when not being able to kill the
\code{\link[processx]{process}} objects. Be aware that you can also
kill Java processes other than the photon application using this function!
}
\details{
A list of running Java tasks is retrieved using \code{ps} (on Linux and MacOS)
or \code{tasklist} (in Windows). Tasks are killed using \code{pkill}
(on Linux and MacOS) or \code{Taskkill} (on Windows).
}
\examples{
if (FALSE) {
purge_java() # should do nothing

# start a new photon instance
dir <- file.path(tempdir(), "photon")
photon <- new_photon(dir, country = "Samoa")
photon$start()

# kill photon using a sledgehammer
purge_java()
}
}
