% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{reverse}
\alias{reverse}
\title{Reverse Geocode Locations With Photon API}
\usage{
reverse(x, y, server = NULL)
}
\arguments{
\item{x}{Longitude(s).}

\item{y}{Latitude(s).}

\item{server}{Select which API is used. The default API is the photon public
API (see Details), but you can use your own photon API.}
}
\value{
A data.frame with the following fields is returned :\cr
x, y, osm_id, osm_type, name, housenumber, street, postcode, city, state,
country, osm_key, osm_value,  lon, lat, msg.\cr
\code{x} and {y} are the original searched coordinates\cr
\code{msg} is filled with the "Not found" text string if no results are found for a
location.
}
\description{
This function reverse geocode locations using the photon API.
}
\details{
Terms of use of the photon public API: \cr
"You can use the API for your
project, but please be fair - extensive usage will be throttled. We do not
guarantee for the availability and usage might be subject of change in the
future. Have fun with photon and make OSM grow!"
}
\examples{
\dontrun{
address <- c("19 rue Michel Bakounine, 29600 Morlaix, France",
             "5 rue Proudhon, 34130 Mauguio France",
             "2 Emma Goldmanweg, Tilburg, Netherlands",
             "36 Strada Panait Israti, Bucarest, Romania")
place <- geocode(address, limit = 1, key = "place")
place2 <- reverse(x = place$lon, y = place$lat)
place2

identical(place[,2:12], place2[,3:13])
}
}
\references{
Photon web site \url{http://photon.komoot.de/}\cr
Photon source code \url{https://github.com/komoot/photon}
}
