% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structured.R
\name{structured}
\alias{structured}
\title{Structured geocoding}
\usage{
structured(
  .data,
  limit = 3,
  lang = "en",
  bbox = NULL,
  osm_tag = NULL,
  layer = NULL,
  locbias = NULL,
  locbias_scale = NULL,
  zoom = NULL,
  progress = interactive()
)
}
\arguments{
\item{.data}{Dataframe or list containing structured information on a place
to geocode. Can contain the columns \code{street}, \code{housenumber},
\code{postcode}, \code{city}, \code{district}, \code{county}, \code{state},
and \code{countrycode}. At least one of these columns must be present in the
dataframe.}

\item{limit}{Number of results to return. Defaults to 3.}

\item{lang}{Language of the results.}

\item{bbox}{Any object that can be parsed by \code{\link[sf]{st_bbox}}.
Results must lie within this bbox.}

\item{osm_tag}{Character string giving an
\href{https://wiki.openstreetmap.org/wiki/Tags}{OSM tag} to filter the
results by. See details.}

\item{layer}{Character string giving a layer to filter the results by.
Can be one of \code{"house"}, \code{"street"}, \code{"locality"},
\code{"district"}, \code{"city"}, \code{"county"}, \code{"state"},
\code{"country"}, or \code{"other"}.}

\item{locbias}{Numeric vector of length 2 or an \code{sfg} point
geometry that specifies a location bias for geocoding. Geocoding results
are biased towards this point. The radius of the bias is controlled through
\code{zoom} and its prominence through \code{location_bias_scale}.}

\item{locbias_scale}{Numeric vector specifying the prominence of
point in \code{location_bias}. Defaults to 0.2.}

\item{zoom}{Numeric vector specifying the radius of the \code{location_bias}.
Corresponds to the zoom level in OpenStreetMap. Defaults to 16.}

\item{progress}{If \code{TRUE}, shows a progress bar for longer queries.}
}
\description{
Geocode a set of place information such as street, house number, or
post code. Structured geocoding is generally more accurate but requires
more information than \link[=geocode]{unstructured geocoding}. Note that
structured geocoding must be specifically enabled when building a Nominatim
database. It is generally not available on komoot's public API and on
pre-built search indices through \code{\link{download_searchindex}}.
}
\details{
Filtering by OpenStreetMap tags follows a distinct syntax explained on
\url{https://github.com/komoot/photon}. In particular:

\itemize{
 \item{Include places with tag: \code{key:value}}
 \item{Exclude places with tag: \code{!key:value}}
 \item{Include places with tag key: \code{key}}
 \item{Include places wih tag value: \code{:value}}
 \item{Exclude places with tag key: \code{!key}}
 \item{Exclude places with tag value: \code{:!value}}
}
}
